+++ packages/server/src/__tests__/express/integration.test.ts
@@ -5,10 +5,10 @@ import http from 'http';
 import {
   ApolloServer,
   ApolloServerOptions,
-  ApolloServerPluginDrainHttpServer,
   BaseContext,
   expressMiddleware,
 } from '../..';
+import { ApolloServerPluginDrainHttpServer } from '../../plugin/drainHttpServer';
 import { urlForHttpServer } from '../../utils/urlForHttpServer';
 import type {
   CreateServerForIntegrationTestsOptions,
