+++ packages/server/src/__tests__/plugin/cacheControl/cacheControlPlugin.test.ts
@@ -1,11 +1,9 @@
 import { GraphQLError } from 'graphql';
 import {
-  ApolloServer,
   ApolloServerPluginCacheControl,
   ApolloServerPluginCacheControlOptions,
-  CacheHint,
-  HTTPGraphQLResponse,
-} from '../../..';
+} from '../../../plugin/cacheControl';
+import { ApolloServer, CacheHint, HTTPGraphQLResponse } from '../../..';
 
 describe('plugin', () => {
   describe('willSendResponse', () => {
