+++ packages/server/src/index.ts
@@ -8,12 +8,5 @@ export {
 
 export { ApolloServer } from './ApolloServer';
 export { expressMiddleware } from './express';
-export { startStandaloneServer } from './standalone';
 
 export * from './externalTypes';
-
-export * from './plugin';
-
-// TODO(AS4): We do need to export usage reporting protobuf for the purpose of
-// gateway, but we really don't want that to be in top level. Look into this as
-// part of #6243.
